+++ src/word.py
@@ -1,0 +1,23 @@
+
+# Data structure for holding data
+class Word():
+    def __init__(self, char, freq = 0, deg = 0):
+        self.freq = freq
+        self.deg = deg
+        self.char = char
+
+    def returnScore(self):
+        return self.deg/self.freq
+
+    def updateOccur(self, phraseLength):
+        self.freq += 1
+        self.deg += phraseLength
+
+    def getChar(self):
+        return self.char
+
+    def updateFreq(self):
+        self.freq += 1
+
+    def getFreq(self):
+        return self.freq
